+++ controller/include/tactile_sensor_impl.h
@@ -40,12 +40,12 @@
 #include <tactile_sensor.h>
 
 namespace ta11_controller {
-TactileSensorBase::TactileSensorBase(ros::NodeHandle& nh, std::vector<std::shared_ptr<float>> forces, bool simulation) : nh_(nh), forces_(forces), sim(simulation){
+TactileSensorBase::TactileSensorBase(ros::NodeHandle& nh, std::vector<std::shared_ptr<double>> forces, bool simulation) : nh_(nh), forces_(forces), sim(simulation){
   for (int i = 0; i < forces_.size(); i++)
     tmp_forces_.push_back(0.0);
 }
 
-TactileSensorSub::TactileSensorSub(ros::NodeHandle& nh, std::vector<std::shared_ptr<float>> forces) : TactileSensorBase(nh, forces, true) {
+TactileSensorSub::TactileSensorSub(ros::NodeHandle& nh, std::vector<std::shared_ptr<double>> forces) : TactileSensorBase(nh, forces, true) {
     sub_ = nh.subscribe("/ta11", 0, &TactileSensorSub::sensor_cb_, this);
     ROS_INFO_STREAM("Registered subscriber for \"/ta11\"");
 }
@@ -62,7 +62,7 @@ void TactileSensorSub::update() {
     }
 }
 
-TactileSensorReal::TactileSensorReal(ros::NodeHandle& nh, std::vector<std::shared_ptr<float>> forces) : TactileSensorBase(nh, forces, false) {}
+TactileSensorReal::TactileSensorReal(ros::NodeHandle& nh, std::vector<std::shared_ptr<double>> forces) : TactileSensorBase(nh, forces, false) {}
 }
 
 #endif  // TA11_CONTROLLER_TACTILE_SENSOR_IMPL_H
